improved the dynamic list: requires now a (Ordered)Dict with key: the actual property of the class and value: a description for the property added flash_message to main page

+++ webinterface/sensors.py
@@ -2,6 +2,7 @@ import os
 
 # web framework
 import cherrypy
+import urllib
 
 # db connection
 from sqlalchemy.ext.declarative import declarative_base
@@ -18,6 +19,8 @@ from mako.lookup import TemplateLookup
 # our stuff
 from tools.db import objects
 from tools import config
+from collections import OrderedDict
+
 
 lookup = TemplateLookup(directories=['templates'], strict_undefined=True)
 
@@ -28,24 +31,33 @@ class SensorsPage(object):
 		return cherrypy.request.db
 	
 	@cherrypy.expose
-	def index(self):
+	def index(self, flash_message=None):
 		tmpl = lookup.get_template("sensors.mako")
-		return tmpl.render(page_title="Sensors")
+		return tmpl.render(page_title="Sensors", flash_message=flash_message)
 		
 	@cherrypy.expose
-	def list(self):
+	def list(self, flash_message=None):
 		tmpl = lookup.get_template("list.mako")
-		
 		sensors = self.db.query(objects.Sensor).all()
 		
+		fields = OrderedDict()
+		fields['id'] = 'ID'
+		fields['name'] = 'Name'
+		fields['description'] = 'Description'
+		fields['gpio_pin'] = 'GPIO Pin'
+		fields['zone_id'] = 'Zone ID'
 		
-		return tmpl.render(data=sensors, page_title="Sensors List")
+		return tmpl.render(data=sensors, page_title="Sensors List", flash_message=flash_message, fields=fields)
 		
 	@cherrypy.expose
-	def delete(self, id):
+	def delete(self, flash_message, id):
 		if id:
 			sens = self.db.query(objects.Sensor).get(id)
-			self.db.delete(sens)
-			self.db.commit()
-		
-		raise cherrypy.InternalRedirect("list")
+			if(sens):
+				self.db.delete(sens)
+				self.db.commit()
+			else:
+				raise cherrypy.HTTPRedirect("list?%s"% urllib.urlencode({'flash_message': 'ID not found!'}))
+				
+			
+		raise cherrypy.HTTPRedirect("list?%s"%urllib.urlencode({'flash_message': 'Sensor deleted!'}))
